Player class now has an UpgradeManager
This handler now has the functions previously held by the Player class.
